refactor: #108 - TabBarList 컴포넌트 경로 수정
