+++ src/components/Header/components/index.ts
@@ -1,0 +1,1 @@
+export { default as TabBarList } from './TabBarList';
