+++ src/components/Header/index.tsx
@@ -2,7 +2,7 @@ import { PARAM_VALUES, SEARCH_VALUES } from '@constants';
 import styled from '@emotion/styled';
 import Icon from '@components/Icon';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
-import TabBarList from './TabBarList';
+import { TabBarList } from './components';
 import { SELECT_OPTION } from './constants';
 import { useSelect } from './hooks';
